Strip out the heart of Printzini
and stuff it into a module
